mdss: Display related changes commit below
While merging to 3.14 kernel this is only display
related changes for commit below
Change ID: I12e6fb6ec8a8481a7eef374cb9316e5ccbc29090

msm: Update Ion and IOMMU APIs to use dma_addr_t

The current APIs for Ion and the custom IOMMU mapping assume that
many parameters are unsigned long. This type is not big enough to
hold physical addresses on LPAE systems. Update the APIs to accomodate
these types. Because we are updating the APIs, all clients need to be
updated as well, either by using the correct type or using a cast
where necessary.

+++ drivers/video/msm/mdss/mdp3_dma.c
@@ -542,7 +542,7 @@ static int mdp3_dmap_update(struct mdp3_dma *dma, void *buf,
 	}
 	spin_lock_irqsave(&dma->dma_lock, flag);
 	MDP3_REG_WRITE(MDP3_REG_DMA_P_IBUF_ADDR, (u32)buf);
-	dma->source_config.buf = buf;
+	dma->source_config.buf = (int)buf;
 	if (dma->output_config.out_sel == MDP3_DMA_OUTPUT_SEL_DSI_CMD) {
 		mdp3_ccs_update(dma);
 		MDP3_REG_WRITE(MDP3_REG_DMA_P_START, 1);
@@ -579,7 +579,7 @@ static int mdp3_dmas_update(struct mdp3_dma *dma, void *buf,
 
 	spin_lock_irqsave(&dma->dma_lock, flag);
 	MDP3_REG_WRITE(MDP3_REG_DMA_S_IBUF_ADDR, (u32)buf);
-	dma->source_config.buf = buf;
+	dma->source_config.buf = (int)buf;
 	if (dma->output_config.out_sel == MDP3_DMA_OUTPUT_SEL_DSI_CMD)
 		MDP3_REG_WRITE(MDP3_REG_DMA_S_START, 1);
